# src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: Mercahub API
  description: API inspirada no Mercado Livre
  version: v1
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
  /item/{id}:
    get:
      tags:
        - Item
      operationId: getItem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /items:
    get:
      tags:
        - Item
      operationId: listItems
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Lista paginada de itens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemSummaryDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ItemDto:
      type: object
      properties:
        id:
          type: string
        site_id:
          type: string
        title:
          type: string
        family_name:
          type: string
        seller_id:
          type: integer
          format: int64
        category_id:
          type: string
        price:
          type: number
          format: double
        base_price:
          type: number
          format: double
        original_price:
          type: number
          format: double
        currency_id:
          type: string
        initial_quantity:
          type: integer
        available_quantity:
          type: integer
        sold_quantity:
          type: integer
        sale_terms:
          type: array
          items:
            $ref: '#/components/schemas/SaleTerm'
        buying_mode:
          type: string
        listing_type_id:
          type: string
        start_time:
          type: string
          format: date-time
        stop_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        expiration_time:
          type: string
          format: date-time
        condition:
          type: string
        permalink:
          type: string
        thumbnail_id:
          type: string
        thumbnail:
          type: string
        pictures:
          type: array
          items:
            $ref: '#/components/schemas/Picture'
        video_id:
          type: string
        accepts_mercadopago:
          type: boolean
        seller_address:
          $ref: '#/components/schemas/SellerAddress'
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/Attribute'
        variations:
          type: array
          items:
            $ref: '#/components/schemas/Variation'
    ItemSummaryDto:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        price:
          type: number
          format: double
        status:
          type: string
        warranty:
          type: string
        catalog_product_id:
          type: string
        domain_id:
          type: string
        date_created:
          type: string
          format: date-time
        last_updated:
          type: string
          format: date-time
        channels:
          type: array
          items:
            type: string
    SaleTerm:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value_id:
          type: string
        value_name:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/SaleTermValue'
        value_type:
          type: string
    SaleTermValue:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Picture:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        secure_url:
          type: string
        size:
          type: string
        max_size:
          type: string
    SellerAddress:
      type: object
      properties:
        address_line:
          type: string
        zip_code:
          type: string
        city:
          $ref: '#/components/schemas/LocationElement'
        state:
          $ref: '#/components/schemas/LocationElement'
        country:
          $ref: '#/components/schemas/LocationElement'
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        id:
          type: integer
          format: int64
    LocationElement:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Attribute:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value_id:
          type: string
        value_name:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/SaleTermValue'
        value_type:
          type: string
    Variation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        price:
          type: number
          format: double
        attribute_combinations:
          type: array
          items:
            $ref: '#/components/schemas/AttributeCombination'
        available_quantity:
          type: integer
        sold_quantity:
          type: integer
        picture_ids:
          type: array
          items:
            type: string
        user_product_id:
          type: string
    AttributeCombination:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        value_id:
          type: string
        value_name:
          type: string
        values:
          type: array
          items:
            $ref: '#/components/schemas/SaleTermValue'
        value_type:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        status:
          type: integer
        cause:
          type: array
          items:
            type: string